added code for update student

+++ src/main/java/com/example/demo/repo/StudentInterface.java
@@ -8,7 +8,7 @@ import com.example.demo.model.Student;
 public interface StudentInterface {
 	public Student save(Student student);
 	public List<Student> getAllStudents();
-	public Optional<Student> getStudentById(Long id);
-	public void deleteStudentById(Long id);
+	public Optional<Student> getStudentById(int id);
+	public void deleteStudentById(int id);
 
 }
