+++ src/main/java/com/example/demo/repo/StudentRepo.java
@@ -1,14 +1,10 @@
 package com.example.demo.repo;
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.demo.model.*;
+import com.example.demo.model.Student;
 @Repository
-public interface StudentRepo extends JpaRepository<Student, Long>{
+public interface StudentRepo extends JpaRepository<Student, Integer>{
